perf: implements recoil in autonomous crawl
