+++ Erlang-Legacy/Assets/Behaviors/RecoilBehavior.asset
@@ -1,0 +1,38 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b23f08d2ae4cba14087c1ed36193d82b, type: 3}
+  m_Name: RecoilBehavior
+  m_EditorClassIdentifier: 
+  mBehaviorSource:
+    behaviorName: Behavior 2
+    behaviorDescription: 
+    mTaskData:
+      types:
+      - BehaviorDesigner.Runtime.Tasks.EntryTask
+      - BehaviorDesigner.Runtime.Tasks.Sequence
+      - BehaviorDesigner.Runtime.Tasks.HasReceivedEvent
+      - Core.Combat.IA.Behavior.Action.Freeze
+      - Core.Combat.IA.Behavior.Action.Recoil
+      - BehaviorDesigner.Runtime.Tasks.Wait
+      parentIndex: ffffffff0000000001000000010000000100000001000000
+      startIndex: 08000000120000001e0000002c000000360000004a000000
+      variableStartIndex: 0000000004000000
+      JSONSerialization: 
+      fieldSerializationData:
+        typeName: []
+        fieldNameHash: f9be55564aaa52567707a10413172f4cf9be55564aaa52567707a1041af3e5d32944782b951203e811c37978c4a37522ae07a096e44d14616fd219c3a5f3b63149269a1caac4f9012944782b951203e811c37978c4a37522ae07a096e44d14616fd219c3a5f3b63149269a1caac4f9018690b4b290adcd7b2944782b951203e811c37978c4a37522ae07a096e44d14616fd219c3a5f3b63149269a1caac4f9019bf0092e94af5f84e59a5c847ebf8c022944782b951203e811c37978c4a37522ae07a096e44d14616fd219c3a5f3b63149269a1caac4f9012944782b951203e811c37978c4a37522ae07a096e44d14616fd219c3a5f3b63149269a1caac4f901114307630a025db95bed59b9884aa8672b36ed36b9e91dd3b2a873290394702930f1bed7cc004d1f2944782b951203e811c37978c4a37522ae07a096e44d14616fd219c3a5f3b63149269a1caac4f901598f7293524ec8e9a339c5e973825867956c6dd48e2bc32adf16c02a81a8551003ac852ffc6adb854d56d8851d9f6b0315ab852f0e6adb855f55d8852f9e6b03
+        startIndex: 000000000100000002000000030000000400000005000000060000000700000008000000090000000a0000000b0000000c0000000d0000000e0000000f000000100000001100000012000000130000001400000015000000160000001700000018000000190000001a0000001b0000001c0000001c0000001d0000001e0000001f00000020000000210000002200000023000000240000002500000026000000270000002700000028000000290000002a0000002b0000002c0000002d0000002e0000002f000000300000003100000032000000330000003400000035000000360000003700000038000000390000003a0000003b0000003c0000003d0000003e0000003e0000003f00000040000000410000004200000042000000430000004400000045000000460000004700000048000000490000004a0000004b0000004c0000004d0000004e0000004f00000050000000500000005100000052000000530000005300000054000000550000005600000056000000570000005800000059000000590000005a0000005b000000
+        dataPosition: 000000002600000035000000360000003e000000620000006d0000006e00000072000000760000007b0000007c0000007d000000850000008500000086000000870000008b0000008f000000930000009b0000009c0000009d000000a5000000a5000000a6000000a7000000ab000000af000000b3000000b7000000c9000000ca000000cb000000d3000000d3000000d4000000d5000000d9000000dd0000000201000002010000080100000c0100001201000013010000140100001c0100001c0100001d0100001e01000022010000260100002a0100003001000031010000320100003a0100003a0100003b0100003c0100004001000044010000680100007301000074010000780100009e010000ad010000ae010000b6010000ba010000be010000bf010000c0010000c8010000c8010000c9010000ca010000ce010000d2010000f6010000f6010000fa0100001d0200001d0200001e0200004202000042020000460200006a0200006a020000
+        unityObjects: []
+        byteData: 4265686176696f7244657369676e65722e52756e74696d652e536861726564566563746f72327265636f696c446972656374696f6e0100000000000000004265686176696f7244657369676e65722e52756e74696d652e536861726564466c6f61747265636f696c5363616c65010000000000000000456e7472790100bb3c06c46888e3c2000000000000000000000100000053657175656e6365010090259bbf52551943000000000000000000000200000002000000486173205265636569766564204576656e740100f735c4c3ca565343000000000000000000004265686176696f7244657369676e65722e52756e74696d652e536861726564537472696e677265636f696c03000000467265657a6501002cdf14c3db76574300000000000000000000040000005265636f696c0100892ef241b9486243000000000000000000004265686176696f7244657369676e65722e52756e74696d652e536861726564466c6f61747265636f696c5363616c6501000000004265686176696f7244657369676e65722e52756e74696d652e536861726564566563746f72327265636f696c446972656374696f6e0100000000000000000500000057616974010051797243d0df5d43000000000000000000004265686176696f7244657369676e65722e52756e74696d652e536861726564466c6f61740000403f4265686176696f7244657369676e65722e52756e74696d652e536861726564426f6f6c004265686176696f7244657369676e65722e52756e74696d652e536861726564466c6f61740000003f4265686176696f7244657369676e65722e52756e74696d652e536861726564466c6f61740000803f
+        byteDataArray: 
+      Version: 1.6.6
