+++ Erlang-Legacy/Assets/Behaviors/RecoilBehavior.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a75c7c3da5a2ab24696c06191a99f388
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
