+++ Erlang-Legacy/Assets/Core/Combat/Destroyable.cs
@@ -41,5 +41,6 @@ namespace Core.Combat
             health.Revive();
             OnRevived?.Invoke();
         }
+
     }
 }
