+++ Erlang-Legacy/Assets/Core/Combat/Hittable.cs
@@ -1,6 +1,5 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using BehaviorDesigner.Runtime;
 using Core.Shared;
 using Core.Utility;
 using DG.Tweening;
@@ -28,12 +27,13 @@ namespace Core.Combat
         [SerializeField] Animator animator;
         [SerializeField] SpriteRenderer sprite;
         [SerializeField] bool hitShakeCamera;
+        public float recoilScale = 0f;
         protected Vector2 baseScale;
         protected Color baseColor;
         protected Material baseMaterial;
         private Color defaultColor = Color.white;
         private Rigidbody2D body => GetComponent<Rigidbody2D>();
-        private Recoil recoil => GetComponent<Recoil>();
+        protected BehaviorTree behaviorTree => GetComponent<BehaviorTree>();
         private List<Tween> tweens = new List<Tween>();
 
         // Start is called before the first frame update
@@ -55,7 +55,23 @@ namespace Core.Combat
 
         private void Recoil(Vector2 direction)
         {
-            recoil?.OnRecoil(direction);
+            Vector2 norm = direction.normalized;
+            if (behaviorTree)
+                NotifyRecoilEvent(norm);
+            else ApplyRecoil(norm);
+        }
+
+        private void ApplyRecoil(Vector2 direction)
+        {
+            Vector2 recoilForce = direction.normalized * recoilScale;
+            body.AddForce(recoilForce, ForceMode2D.Impulse);
+        }
+
+        private void NotifyRecoilEvent(Vector2 direction)
+        {
+            behaviorTree.GetVariable("recoilDirection")?.SetValue(direction);
+            behaviorTree.GetVariable("recoilScale")?.SetValue(recoilScale);
+            behaviorTree.SendEvent("recoil");
         }
 
         private void ReactHit()
