+++ Erlang-Legacy/Assets/Core/IA/Behavior/Task/Action/Freeze.cs
@@ -1,0 +1,19 @@
+using BehaviorDesigner.Runtime;
+using BehaviorDesigner.Runtime.Tasks;
+using UnityEngine;
+
+namespace Core.Combat.IA.Behavior.Action
+{
+    public class Freeze : EnemyAction
+    {
+        public override void OnStart()
+        {
+            body.velocity = Vector2.zero;
+        }
+
+        public override TaskStatus OnUpdate()
+        {
+            return TaskStatus.Success;
+        }
+    }
+}
