+++ Erlang-Legacy/Assets/Core/IA/Behavior/Task/Action/Freeze.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 65569828d61d76945b9f7ef2f392272a
+guid: 4999dc7649272914b99fad0a8b717573
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
